chore: extract toJSXExpression to own util file
