+++ codemods/clamshells.js
@@ -7,14 +7,8 @@ const {
   removeChild,
   setAttribute,
 } = require('./utils/mdxast');
-const {
-  stringify,
-  mdxAttribute,
-  mdxValueExpression,
-  mdxSpanElement,
-  mdxBlockElement,
-} = require('./utils/mdxast-builder');
-const { root, text } = require('mdast-builder');
+const { stringify, mdxValueExpression } = require('./utils/mdxast-builder');
+const toJSXExpression = require('./utils/to-jsx-expression');
 
 const clamshells = () => (tree, file) => {
   visit(
@@ -60,51 +54,6 @@ const clamshells = () => (tree, file) => {
       );
     }
   );
-};
-
-const toJSXExpression = (node, file) => {
-  const children = transformChildren(node, file);
-  const tree = root(
-    children.length === 1 ? children : [mdxSpanElement(null, [], children)]
-  );
-
-  console.dir(tree, { depth: null });
-
-  return tree;
-};
-
-const transformChildren = (node, file) => {
-  return node.children
-    .flatMap((child) => {
-      const transformer = TRANSFORMERS[child.type];
-
-      if (!transformer) {
-        file.message(
-          `Converting to JSX expression of unknown type: '${child.type}'`,
-          child.position.start,
-          'clamshells'
-        );
-        return;
-      }
-
-      return transformer(child, file);
-    })
-    .filter(Boolean);
-};
-
-const TRANSFORMERS = {
-  paragraph: transformChildren,
-  inlineCode: (node) => mdxSpanElement('code', [], [text(node.value)]),
-  text: (node) => node,
-  link: (node) => {
-    const isRelative = node.url.startsWith('http');
-
-    return mdxSpanElement(
-      isRelative ? 'Link' : 'a',
-      [mdxAttribute(isRelative ? 'to' : 'href', node.url)],
-      node.children
-    );
-  },
 };
 
 module.exports = clamshells;
